Generate a page for each item of content
Use the existing pages for runnable code and basic (minus listing for
the moment) and just dump the raw data into the page for text.

This also required updating the DFS disc reading code to handle both
sides correctly, which means that dconv is no-longer required.

Also ensure that the same page isn't created twice with different
content by checking each entry against a hash as it is created.