Add issue 49 to the conversion
As the menu was overhauled for issue 50 there needs to be separate code
to handle the different format for issue 49 and below.  To handle this
cleanly, move the generic DATA line reading code into EBS::Disc, the
specific issue 50+ code into EBS::Disc50To66, and write new code to
read the data correctly from the issue 49 menu.