Correct behaviour of double height Mode 7 chars
Currently, the Mode 7 double height text doesn't work correctly in cases
where there is a character paired with a space (e.g. a space above a
full-stop), and where double height sections are mis-aligned (such as in
BASIC listings).

To resolve this, instead of storing the previous line's values, instead
store and use an array of the character mode in the row above:
standard, upper double or lower double, and base the upper / lower half
logic on these values.