Support a menu entry having multiple file paths
Issues 28 to 49 can have multiple files for a menu entry (I believe just
for Mode 7 text, due to the length limit), so re-work the file path and
contents handling to allow for this and split file paths on the @ sign
in the issue 28 to 49 disc reader.

Also, some of the secondary paths are also the primary path of a
different menu item, so post-process the paths to remove these to
prevent duplicated content.