Add issue 29 to the conversion
ASCII value 155 needs to be mapped to ' ' in the Mode 7 conversion and
the parsing of PROCla needs to check where the line number value starts
and ignore unused lines.

Also add CSS styles for the <hr> between linked Mode 7 pages as this is
the first issue that this has appeared.