Replace BASIC LDPIC loaders with the loaded images
Implement the ability to override automatically converted menu data with
YAML files in the _data directory and use this to replace simple BASIC
loaders which just call LDPIC with the images themselves.

Also add the facility for captioning images and use this to include any
relevant text from the BASIC loaders, as well as for the Issue 7 images.