Replace screen dump loaders with their images
Replace simple BASIC loaders which just *LOAD screen dumps with the
images themselves.  This required adding a 'modes' parameter for setting
the mode (and optionally the colour mappings), as this can't easily be
inferred from the screen dumps.

Also allow the 'offsets' parameter to be used for all file types, as
some of the screen dumps also contain other data.