Emulate & list BASIC progs in ARCer v3.0 archives
Load archives generated by ARCer version 3.0 (used in issues 18 and 19)
so that BASIC programs within them can be listed in the same way as
those stored directly in the disc images.

To enable emulation of programs within the archives, build .ssd disc
images containing the archived files and load these instead of the issue
disc image into jsbeeb when emulating an archived program.