Extend archive unpacking to 'ARC' format in 8BS22
Issue 22 has a number of archives in 'ARC' format (which originated as a
DOS archiving tool) which compresses the files contained within it as
well as combining them into a single file.

To avoid having to re-implement the decompression routines, use a couple
of the original ARC source files (and headers), as these were
re-licensed under the GPL in 2003.  Once these were hooked up to some
code to handle setting up the input and output buffers and variables,
these only needed very minimal modification to #ifdef out the unused
parts and call rb_fatal if there were any problems.

Also fix a bug in the disc image generation code - the top two bits of
the start sector were being written to the wrong bits in the relevant
byte.