Improve About & Menu page layout on small screens
Switch the header to be scrollable instead of always fixed in place at
the top of the screen as this uses valuable mobile screen space.
Abbreviate 8-Bit Software to 8BS on smaller screens and only set a
max-width so that it can shrink to fit.

Change the #content width to a max-width to allow the About and Menu
page content to fit to the available screen space.