Make header HTML and CSS cleaner
Simplify the header HTML and make it use more semantic elements (e.g.
header, h1, nav) instead of just spans and divs.  Switch to an unordered
list for the items inside the nav element to structure them better.

Update and clean up the header CSS so that it works with the new HTML
structure.  Also move the navigation separators ">" from the HTML to the
stylesheet as these are presentation rather than content.