Start splitting ContentItem object from MenuEntry
A lot of code relating to output content handling has ended up in the
MenuEntry class.  Start moving this in to a new ContentItem class which
is passed to the content pages instead of the menu entry.

ContentItem takes a list of files instead of paths which is cleaner and
also flagged up a couple of mistakes in the YAML data files.  Fix these
and also add a check in the DfsDisc file() method to catch these issues
more easily in future.