Combine the BBCFile and ArchiveFile classes
As these serve very similar purposes, reduce code duplication by folding
the ArchiveFile class into the BBCFile class.  This required changing
the DfsDisc class to extract the file content when the catalogue is read
rather than when the content is requested but this doesn't appear to
have an adverse effect on build time.