Start moving BASIC parsing out into one class
Currently there are several different classes which parse BASIC files
for different reasons which has lead to a fair amount of duplicated
code.  Start moving towards centralising this all in one place by
pulling the DATA line parsing out of EBS::Disc into a new BasicFile
class.

As there are so many BASIC files to read and convert and the format of
the files lends itself to more low-level parsing, implement the new
BasicFile class in C.

Also add rspec examples for the initial behaviour of the class and add
running rspec to Travis CI.