Use new BasicFile class to generate HTML & bug fix
Replace use of the original basic_to_text filter with a new .to_html
method on the BasicFile class which joins the hash of lines together and
passes the result through the text to memory and memory to html
conversion functions.

Also add/extend examples and fix BasicFile handling of DATA lines, lines
with leading spaces, control characters in REMarks and EOF handling so
that the output matches BASIC IV as closely as possible.