Add 'proof' task to check html with HTMLProofer
Add a new task to rake which builds the site to a temporary folder and
then runs HTMLProofer on the result.

Fix the issues that this found - ampersands aren't strictly necessary to
escape under HTML5 (it doesn't hurt though) but the duplicate menu
numbering in 8BS20 definitely was an issue that hadn't been spotted
previously.

Add running the new proof task to the Travis CI build, excluding the
bundle and moving the libbeebimage build up a level.