Convert all BASIC files present in the disc images
Some BASIC programs are mentioned in the text of the magazine but are
not accessible from the menus themselves which means that they aren't
available to run directly or as browsable listings.

To remedy this issue add a new 'File list' page which lists the
converted files by filename and add any unreferenced files which parse
as valid BASIC to the conversion.

This also required escaping some more values in the liquid templates due
to the presence of special characters in the filenames.