Add a bare-bones class to read BASIC data files
Copy some of the EBS::Archive class into a new BBC:BasicDataFile class
and add a couple of very simple specs for validation.  This code is more
generally useful than just in archive files alone as it could be used
for parsing some information out of data files on some of the discs.

Add a new spec helper method for easily building a BBCFile from a
content string to allow adding tests for trivial file content without
having to add another content file.

Also take the opportunity to add some methods to BBCFile to allow
convenient sequential access to parts of the binary content string
without having to convert the whole string to an array first.  This also
allows convenient use of unpack for wrangling the values.  The methods
added are equivalent to those for array: shift and empty?, the length
method was also updated to be correct for the current position in the
string.