Process files contained within archives separately
Currently, only one file contained within an archive is processed in the
normal way and only if it is referenced by a menu entry.  Perform some
fairly heavy-duty refactoring of the ContentMapper class so that files
nested within an archive file can all be processed in a normal way and
output to paths nested below the path for the archive.  The archive
itself is presented as a list of the files that it contains.

To allow menu entries to link to files within archives, add an extra
dimension to the files array so that each file can be specified as
multiple levels of files.  This also needed quite a lot of small changes
to the classes which handle these arrays.

Also take the opportunity to refactor the archive reading code into
BBC::Arc*File classes so that they can detect archive files not linked
to by any menu entries.  Implement the whole of the ArcFile class in C
for tidiness as it uses the C routines directly from arc itself, but
fold this code back into the bbc_native extension instead of having it
in a separate extension.  Also add examples for the standard behaviour
of these classes.

To handle a couple of odd cases (setting load and exec addresses for a
file from an arc archive and merging two half archives into one), add a
new optional ':tweaks' key to the YAML data files and extend the BBCFile
and DFSDisc classes to pass around and act on the data.

Also bump the Ruby version used in GitHub Actions to 2.7 so that we can
use the convenient flat_map method on arrays.