Use Cppcheck to lint the c extension source
Add a cppcheck rake task (and call it from the lint task) which runs
Cppcheck against the .c files in the _ext/BBC/ folder.  Files in the arc
folder aren't included as they are mostly unchanged from the arc source.

The version of Cppcheck in Ubuntu 18.04 (1.82) contains a bug that
causes the error exit code to be returned even if an error is suppressed
(https://trac.cppcheck.net/ticket/7792), so set the lint job to run
under Ubuntu 20.04 for the time being.

Also take the opportunity to switch GCC into standard c99 mode so
Cppcheck can be configured to check against the same standard.