Add overcommit pre-commit hook for running specs
Add a custom pre-commit hook to validate that the specs all pass.  As
some of them depend on the c extension being built, add a new rspec rake
task which depends on the compile task and run this instead of calling
rspec directly.

Although an RSpec hook is included by default in overcommit, this is a
pre-push hook instead of pre-commit.  The compilation and specs are
quick to run and I prefer for each commit to compile and have passing
specs.  The hook can always be skipped when making WIP commits during
development.