Move author copyright to individual files
Instead of trying to handle the list of authors separately to the file
copyright, remove AUTHORS.md and change the copyright in all the files
to my name instead (as I'm the only author thus far).  Any other future
contributors can add their own copyright notices directly to the files
they contribute to.

Adjust the Copyright overcommit hook to be able to handle multiple
copyright notices correctly and check the commit author's notice is
present and updated.

Also take the opportunity to reduce the size of the file headers by
removing the GPL boilerplate from each file and replacing it with a
SPDX-License-Identifier header line.