Add an action to diff the generated site for PRs
Add an action which runs when a PR is raised and produces a diff between
the output generated by the target branch and the PR which is then added
as a comment.  This should allow easier merging of dependency PRs (which
shouldn't generally change the output) without having to perform the
comparison manually.

This is a bit involved to set up as for security reasons workflow jobs
triggered from the pull_request event only have read-only permissions
(so can't actually add the comment).  Work around this in the manner
recommended by GitHub by saving the result of the diff and the PR number
to an asset and using them in a second workflow triggered on completion
of the first.