Resize the emulator iframe more dynamically
Keep the iframe width at 100% but set a max-width which makes the
content roughly the same size as in the Mode 7 converted pages.

Set the iframe height to roughly the amount of remaining vertical space
but with a limit to keep things fairly sensible.  This may be better
implemented with JavaScript once other changes have been made to the
emulate pages but should be an improvement for now.

This gives better results on mobile and should make better use of larger
screens.