Copy captions from menu entry to content item
Although image captions were defined in the data files, these were not
being displayed in the image content pages.  This was due to the
content_image layout checking the ContentItem for a defined 'captions'
property but this never being pulled through from the menu item.

It looks like this issue was introduced in 99449294db as part of
splitting off ContentItem from MenuEntry.