Change Mode 7 flashing text from JavaScript to CSS
According to caniuse.com, just under 97% of users have browsers which
support CSS animations and keyframes, so take the opportunity to switch
out the JavaScript used to animate any flashing text in Mode 7 for a CSS
equivalent that will work without scripts enabled.

Also, correct the flash timings (which were previously a guess based on
emulator behaviour).  Details posted on stardot at
https://stardot.org.uk/forums/viewtopic.php?p=306475#p306475 (which were
obtained by reverse engineering the SAA5050 Teletext chip) below:

"This is possibly getting a bit too exact, but I think the SAA5050
datasheet is wrong about the frequency of the flash. The Flash Counter
is 6 bits and counts from 0 to 63. There is no regular reset for it, so
it just counts up to 63 then back to 0. The flash is OFF for the first
16 values, and ON for the other 48 values. That is a 1:3 OFF/ON ratio,
which agrees with the spec, but if you calculate the frequency of the
flash, I get 0.78125 Hz and not 0.75 Hz. The counter increments in
response to the external DEW signal, so once per field, i.e. 50 times a
second."

By my calculations, this means that the whole cycle takes 64/50 = 1.28
seconds and is off for 16/64 = 25% of the time, so use these timings in
the CSS.