Remove Mode 7 scroller headers in a separate class
Instead of handling this case within ContentItem, move the calculation
and removal of the scrolling text headers to a new class called
Dispmo7File (after the most recent name of the conversion program).
Make the parsing much more robust with various sanity checks and add
specs for the various failure cases as well as successful extraction of
the data.

As some of the checks require the execution address of the file, extend
DfsDisc to also fetch this value from the catalogue.

Make the Dispmo7File class a child of a new Mode7File class as there are
other Mode 7 screen formats which can inherit from it in future.

During implementation of the above, discovered that the header of
$.Soliad on side 2 of 8BS30 has an invalid header (both the start/end
addresses and the loader itself are just repeated 0x20 bytes), so add
manual extraction offsets for this file.