Make AbzFile parsing stricter in 25-line mode
Check the remainder of the data between the end of line 25 and the 1KiB
boundary to ensure that it contains some non-printable characters.  If
this data is all printable too it is highly unlikely that the file being
parsed is in ABZ format.  Otherwise, some graphics mode screen dumps
(e.g. those with most bits set) could be incorrectly interpreted as
being ABZ format.