Treat the 'TELEMAG' file in 8BS16-1 as an archive
This adds a new TelemagFile class which inherits from ArchiveFile and
sequentially names the different article 'files' as A, B, C, etc.
Although this file format only appears once, more importantly it is the
only instance of different byte ranges in a single file being referenced
by separate menu items and splitting the content out like this will make
future refactoring much easier.

As the format only occurs once add a new key to the 'tweaks' hash for
files which allows specifying the class name used for parsing instead of
adding the class to the list tried for every file.  Also set this for
the child article files so they they are correctly handled as MODE 7
screen data.