Replace content item offsets with file byteranges
Move definition of byte ranges to select out of files from menu
entries/menu overrides to the files themselves (via the tweaks hash).
This is a tidier and more flexible way to define this as well as being
much cleaner from an implementation perspective.  This did however
require moving the menu item overrides for issue 28+ from the top level
of the data files to under a :menus key.