Example for all Mode 7 characters in alpha mode
Add an example to test the mapping of all non-control characters in Mode
7 alpha mode.  Some of the values don't appear to be correct but the
glyphs at those code points in the Mode 7 font are the correct ones to
display (e.g. [ actually displays as ←).

To allow rspec to compare the expected and actual result, correctly
define the result as UTF-8.  This doesn't affect the generated site as
the strings already comprise of UTF-8 bytes but declares them as such.

This revealed two bugs in the Mode 7 conversion, the mapping of
0x7E/0xFE to ~ instead of ÷ and of 0x7F/0xFF to the same entirely filled
character as in graphics mode instead of the correct character (which is
filled except for a border).