Extend the Mode 7 reserved characters example
Although this tested that all of the characters were mapped to space, it
did not test that the characters had no side effects.  This missed that
0x0A was incorrectly being treated the same as 0x0D (double height), so
add full stops at the start, end and between each character.

Also correct the mapping of 0x0A in the conversion so that the example
passes again.