Handle EDIT converter mode of old Mode 7 scroller
The Mode 7 text displayed at /8BS19/content/0-c-texted/0-disinfo/
doesn't match what is displayed when the original file is run as it has
the 'EDIT converter' feature of the early (Mo7dis) loader enabled which
remaps various control chars before displaying.

Add an example to the spec to handle this behaviour and implement
detecting when the conversion is enabled and performing it in the
Dispmo7File class.