Add held graphics related examples & fix handling
Add examples for all of the beebjit 'teletest' Mode 7 test cases which
feature held graphics.  These were written by scarybeasts (Chris Evans)
& Tom Seddon and cover some interesting corner cases.  I have changed
uses of colour to white in the tests where it isn't central to the case
so that the output being tested doesn't contain as much HTML.

Fix the behaviour of the Mode 7 conversion so that all of the examples
pass and add a couple of examples of my own to cover other incorrect
behaviours caused while doing so.