Add missing second double teletest & fix behaviour
Add the MISSING SECOND DOUBLE teletest by scarybeasts (Chris Evans) as a
Mode 7 example (but changed to white for clarity).  This checks that
normal height characters are not displayed on the line below the upper
half of double height characters.  Adjust the conversion to implement
this behaviour and correct the handling of double upper/lower half state
handling to make it per-line instead of per-character as it was
previously.

Also fix some of the existing examples to be aligned with the correct
behaviour/have better coverage and an example for the blank characters
behaviour being reset correctly.