Mode 7: Optimise out repeated bg or flash spans
Currently, if a text or graphics colour code is used repeatedly only one
span will be opened but if flash or new background is used repeatedly
there will be a succession of spans with identical styles generated.

Fix this behaviour by storing and then comparing the span colour and
flashing style states with the current values so that a new span is only
opened when required.

Also add examples for the correct behaviour for colour, background
colour and flashing.