Skip Mode 7 colour & flash changes for spaces
Delay style changes for text colour and flashing state changes until a
visible (e.g. non-space) character is about to be written and allow
style spans to remain open at the end of a line.  This optimises out
quite a large amount of the generated HTML (roughly 5% of the entire
site) which was being repeated over paragraphs of single-colour text and
styles which didn't visibly affect the output.

Add examples of the correct behaviour for invisible style changes and
styles wrapping on to the next line and update existing examples to
reflect the new locations of <span> tags in the output.