Set all chars to spaces in Mode 7 when fg==bg
When the foreground and background colours are the same in the Mode 7
conversion, replace all non-space characters with spaces because this is
not data that would ever have been seen by the user (so shouldn't be
visible to search engines or by highlighting the text).  Also treat the
effective foreground colour as white so that sections of background
colour filled with spaces are written as <span style=bX> instead of
<span style="fX bX">

Add an example for the character replacement with spaces and update
other examples for the foreground colour style omission as well as
adding an example to cover a held character issue discovered during
implementation.

Also reduce the forecolour var scope to within the character processing
loop as this is set at the start of each iteration.