Mode 7: Include concealed content (but hidden)
Instead of replacing all content with spaces when Conceal Display is
active, allow it through to the generated page but make it hidden with a
new 'conceal' style (which sets color: transparent).  This allows the
user to reveal it by highlighting the hidden text.

This doesn't really matter either way as it appears that concealed
content is only found in /8BS36/content/0-gen1/ (and looks to have been
included there by accident) but it is nice to include it for
consistency.

Add examples for conceal and reset at line end as well as span style
handling and update one existing teletest example to match the new
behaviour.