Rewrite and expand 'Why?' section of about page
When re-reading this section more recently, I realised that the current
'Why?' section could be potentially read as being indirectly negative
about the converted magazines on the official 8BS site.  As this isn't
my intention, rewrite and expand this section to explain the original
motivation (without making any external comparisons).  Also mention when
jsbeeb was originally released as this allowed a significant improvement
to the site.