Update Ruby to 3.2.x
"Dependabot failed to update your dependencies because there was an
error resolving your Ruby dependency files."

Dependabot appears to have started considering the current Ruby version
as part of the dependency constraints (or perhaps changed how version
numbers not including an additional specifier after the minor number are
considered).  As one of HTMLProofer's dependencies depends on Ruby 3.1.1
or above and we're specifying Ruby 3.1 this causes Dependabot to error.

Increase the Ruby version to 3.2 so that we are definitely using a
version greater than 3.1.1 and therefore resolve the Dependabot errors.