Add issue 67 to the conversion
There was an extra issue of the 8BS magazine released in 2013 which I
recently became aware of.  Add this to the conversion too as it seems a
shame to be missing an issue!

Apart from adding the disc image and changing instances of 66 to 67 in
the disc classes, also adjust the date fetching code for issue 50 and
above so that the value can be overridden.  This allows showing the
exact date of release (19/Apr/2013) instead of just the month and year
(there isn't space in the 8BS menu layout for the day when a four digit
year is used).

Also add menu tweaks to link the images directly from the menu.